fix: rename faceboot SSO component to facebook SSO
